+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 ---------
 
 Unreleased
+- remove Python 2.x imports ([#96](https://github.com/stchris/untangle/pull/96))
 
 1.2.1
 - (SECURITY) Use [defusedxml](https://github.com/tiran/defusedxml) to prevent XML SAX vulnerabilities ([#94](https://github.com/stchris/untangle/pull/94))